Rename crunchbase geocoding task.
